# docs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Flight Booking System API
  description: |
    A comprehensive flight booking system API built with Next.js and Supabase.
    
    ## Features
    - Flight search with advanced filters
    - User booking management
    - Admin flight management
    - Real-time notifications
    - Seat selection
    - Email confirmations
    
    ## Authentication
    This API uses JWT-based authentication via Supabase. Include the access token in the Authorization header as `Bearer <token>`.
    
    ## Rate Limiting
    - General endpoints: 100 requests/minute per IP
    - Search endpoints: 20 requests/minute per IP
    - Admin endpoints: 200 requests/minute per admin
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@yourflightbooking.com
  license:
    name: MIT
servers:
  - url: https://api.yourflightbooking.com
    description: Production server
  - url: https://api-test.yourflightbooking.com
    description: Test server
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/flights/{id}/seats:
    get:
      tags:
        - User
      summary: Get available seats
      description: Retrieve available seats for a flight
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Flight ID
      responses:
        '200':
          description: Available seats information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT access token
  schemas:
...
    SeatsResponse:
      type: object
      properties:
        available_seats:
          type: integer
        seat_map:
          type: array
          items:
            type: object
          description: Seat configuration data
        occupied_seats:
          type: array
          items:
            type: string
          description: List of occupied seat numbers
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code (optional)
        details:
          type: object
          description: Additional error details (optional)
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not authenticated"
...